Probamos Las cosas
Probé inscribir un alumno a tres materias y anduvo bien,hubo que cambiar count materias para que devuelva long porque la query devolvía un large int,agregamos unos TODO y habría que evitar bobby tables

+++ WizardUniversityOfNikolaus/Contracts/Repositories/IAlumnoRepository.cs
@@ -19,7 +19,7 @@ namespace Contracts.Repositories
 
         public Task AlumnosDeProfesorAsync(int idProfesor);
 
-        public Task<int> CountMateriasAsync(int idAlumno);
+        public Task<long> CountMateriasAsync(int idAlumno);
 
         public Task<bool> InsertAlumnoEnMateriaAsync(int idAlumno, int idMateria);
